+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+### Added
+
+### Fixed
+
+## [0.3.2-beta] - 2024-05-29
+
 - Fix Delivery Error message (#68)
 - Race condition on log file writing (#67)